Moved DB.GeometryObject extension methods to External.DB.Extensions namespace.
