+++ src/RhinoInside.Revit.GH/Components/Element/DirectShape/ByGeometry.cs
@@ -2,9 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
-using Autodesk.Revit.DB;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Types;
+using RhinoInside.Revit.External.DB.Extensions;
+using DB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Components
 {
@@ -30,21 +31,21 @@ namespace RhinoInside.Revit.GH.Components
 
     void ReconstructDirectShapeByGeometry
     (
-      Document doc,
-      ref Autodesk.Revit.DB.Element element,
+      DB.Document doc,
+      ref DB.Element element,
 
       [Optional] string name,
-      Optional<Autodesk.Revit.DB.Category> category,
+      Optional<DB.Category> category,
       IList<IGH_GeometricGoo> geometry,
-      [Optional] IList<Autodesk.Revit.DB.Material> material
+      [Optional] IList<DB.Material> material
     )
     {
       var scaleFactor = 1.0 / Revit.ModelUnits;
 
-      SolveOptionalCategory(ref category, doc, BuiltInCategory.OST_GenericModel, nameof(geometry));
+      SolveOptionalCategory(ref category, doc, DB.BuiltInCategory.OST_GenericModel, nameof(geometry));
 
-      if (element is DirectShape ds && ds.Category.Id == category.Value.Id) { }
-      else ds = DirectShape.CreateElement(doc, category.Value.Id);
+      if (element is DB.DirectShape ds && ds.Category.Id == category.Value.Id) { }
+      else ds = DB.DirectShape.CreateElement(doc, category.Value.Id);
 
       using (var ga = Convert.Context.Push())
       {
@@ -64,7 +65,7 @@ namespace RhinoInside.Revit.GH.Components
                                          material[materialIndex++]?.Id :
                                          material[materialCount - 1]?.Id
                                         ) ??
-                                        ElementId.InvalidElementId;
+                                        DB.ElementId.InvalidElementId;
                       }
 
                       return x.ToHostMultiple(scaleFactor);
@@ -118,21 +119,21 @@ namespace RhinoInside.Revit.GH.Components
 
     void ReconstructDirectShapeTypeByGeometry
     (
-      Document doc,
-      ref Autodesk.Revit.DB.ElementType elementType,
+      DB.Document doc,
+      ref DB.ElementType elementType,
 
       string name,
-      Optional<Autodesk.Revit.DB.Category> category,
+      Optional<DB.Category> category,
       IList<IGH_GeometricGoo> geometry,
-      [Optional] IList<Autodesk.Revit.DB.Material> material
+      [Optional] IList<DB.Material> material
     )
     {
       var scaleFactor = 1.0 / Revit.ModelUnits;
 
-      SolveOptionalCategory(ref category, doc, BuiltInCategory.OST_GenericModel, nameof(geometry));
+      SolveOptionalCategory(ref category, doc, DB.BuiltInCategory.OST_GenericModel, nameof(geometry));
 
-      if (elementType is DirectShapeType directShapeType && directShapeType.Category.Id == category.Value.Id) { }
-      else directShapeType = DirectShapeType.Create(doc, name, category.Value.Id);
+      if (elementType is DB.DirectShapeType directShapeType && directShapeType.Category.Id == category.Value.Id) { }
+      else directShapeType = DB.DirectShapeType.Create(doc, name, category.Value.Id);
 
       using (var ga = Convert.Context.Push())
       {
@@ -151,7 +152,7 @@ namespace RhinoInside.Revit.GH.Components
                                          material[materialIndex++]?.Id :
                                          material[materialCount - 1]?.Id
                                         ) ??
-                                        ElementId.InvalidElementId;
+                                        DB.ElementId.InvalidElementId;
                       }
 
                       return x.ToHostMultiple(scaleFactor);
@@ -205,8 +206,8 @@ namespace RhinoInside.Revit.GH.Components
 
     void ReconstructDirectShapeByLocation
     (
-      Document doc,
-      ref Autodesk.Revit.DB.Element element,
+      DB.Document doc,
+      ref DB.Element element,
 
       [Description("Location where to place the element. Point or plane is accepted.")]
       Rhino.Geometry.Plane location,
@@ -215,24 +216,24 @@ namespace RhinoInside.Revit.GH.Components
     {
       var scaleFactor = 1.0 / Revit.ModelUnits;
 
-      if (element is DirectShape ds && ds.Category.Id == type.Category.Id) { }
-      else ds = DirectShape.CreateElement(doc, type.Category.Id);
+      if (element is DB.DirectShape ds && ds.Category.Id == type.Category.Id) { }
+      else ds = DB.DirectShape.CreateElement(doc, type.Category.Id);
 
       if(ds.TypeId != type.Id)
         ds.SetTypeId(type.Id);
 
-      var library = DirectShapeLibrary.GetDirectShapeLibrary(doc);
+      var library = DB.DirectShapeLibrary.GetDirectShapeLibrary(doc);
       if (!library.ContainsType(type.UniqueId))
         library.AddDefinitionType(type.UniqueId, type.Id);
 
       var transform = Rhino.Geometry.Transform.PlaneToPlane(Rhino.Geometry.Plane.WorldXY, location.ChangeUnits(scaleFactor)).ToHost();
-      ds.SetShape(DirectShape.CreateGeometryInstance(doc, type.UniqueId, transform));
+      ds.SetShape(DB.DirectShape.CreateGeometryInstance(doc, type.UniqueId, transform));
 
-      var parametersMask = new BuiltInParameter[]
+      var parametersMask = new DB.BuiltInParameter[]
       {
-        BuiltInParameter.ELEM_FAMILY_AND_TYPE_PARAM,
-        BuiltInParameter.ELEM_FAMILY_PARAM,
-        BuiltInParameter.ELEM_TYPE_PARAM
+        DB.BuiltInParameter.ELEM_FAMILY_AND_TYPE_PARAM,
+        DB.BuiltInParameter.ELEM_FAMILY_PARAM,
+        DB.BuiltInParameter.ELEM_TYPE_PARAM
       };
 
       ReplaceElement(ref element, ds, parametersMask);
