+++ src/RhinoInside.Revit/External/DB/Extensions/Curve.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections.Generic;
 using Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.External.DB.Extensions
@@ -7,6 +9,35 @@ namespace RhinoInside.Revit.External.DB.Extensions
     public static bool IsSameKindAs(this Curve self, Curve other)
     {
       return self.IsBound == other.IsBound && self.GetType() == other.GetType();
+    }
+
+    public static IEnumerable<Curve> ToBoundedCurves(this Curve curve)
+    {
+      switch (curve)
+      {
+        case Arc arc:
+          if (!arc.IsBound)
+          {
+            yield return Arc.Create(arc.Center, arc.Radius, 0.0, Math.PI, arc.XDirection, arc.YDirection);
+            yield return Arc.Create(arc.Center, arc.Radius, Math.PI, Math.PI * 2.0, arc.XDirection, arc.YDirection);
+          }
+          else yield return arc;
+          yield break;
+        case Ellipse ellipse:
+          if (!ellipse.IsBound)
+          {
+#if REVIT_2018
+            yield return Ellipse.CreateCurve(ellipse.Center, ellipse.RadiusX, ellipse.RadiusY, ellipse.XDirection, ellipse.YDirection, 0.0, Math.PI);
+            yield return Ellipse.CreateCurve(ellipse.Center, ellipse.RadiusX, ellipse.RadiusY, ellipse.XDirection, ellipse.YDirection, Math.PI, Math.PI * 2.0);
+#else
+            yield return Ellipse.Create(ellipse.Center, ellipse.RadiusX, ellipse.RadiusY, ellipse.XDirection, ellipse.YDirection, 0.0, Math.PI);
+            yield return Ellipse.Create(ellipse.Center, ellipse.RadiusX, ellipse.RadiusY, ellipse.XDirection, ellipse.YDirection, Math.PI, Math.PI * 2.0);
+#endif
+          }
+          else yield return ellipse;
+          yield break;
+        case Curve c: yield return c; yield break;
+      }
     }
   }
 }
