+++ src/RhinoInside.Revit/External/DB/Extensions/GeometryObject.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Collections.Generic;
+using Autodesk.Revit.DB;
+
+namespace RhinoInside.Revit.External.DB.Extensions
+{
+  public static class GeometryObjectExtension
+  {
+    public static IEnumerable<GeometryObject> ToDirectShapeGeometry(this GeometryObject geometry)
+    {
+      switch (geometry)
+      {
+        case Point p: yield return p; yield break;
+        case Curve c:
+          foreach (var unbounded in c.ToBoundedCurves())
+            yield return unbounded;
+          yield break;
+        case Solid s: yield return s; yield break;
+        case Mesh m: yield return m; yield break;
+        case GeometryInstance i: yield return i; yield break;
+        default: throw new ArgumentException("DirectShape only supports Point, Curve, Solid, Mesh and GeometryInstance.");
+      }
+    }
+  }
+}
